QR_maker/requirements.txt
myqr
Tkinter
pillow
python-imaging-tk
